pyproject.toml
[project]
name = "podcast-mcp"
version = "0.1.0"
description = "MCP server for generating podcast dialogues using Coqui TTS"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "mcp>=1.0.0",
    "coqui-tts>=0.27.2",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "numpy<2.0.0",
    "toon-python>=0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/podcast_mcp"]

[project.scripts]
podcast-mcp = "podcast_mcp.server:main"
